elf: Move Elf64Hdr to separate module
In the effort of splitting up the ELF parser code, move the Elf64Hdr
implementation to its own module.
